Restructure parser and compiler
Parser takes the tokens and produces a chunk, as before. But now things are rearranged as a stack of
FunctionCompilers. The job of a function compiler is to compile the chunk associated to "the current
function". Right now the current function is just the top level of the program. But soon we'll start
defining functions inside our functions.